refactor(topology): isolate pseudomanifold contracts
